- Optimize command - Add RecipeWizard to conveniently generate recipes.

+++ common/src/main/java/cool/muyucloud/croparia/recipe/RitualStructure.java
@@ -92,7 +92,7 @@ public class RitualStructure implements Recipe<RitualStructureContainer> {
         return this.patterns.getFirst().maxZ();
     }
 
-    public @Nullable BlockState matchTransformed(BlockPos origin, Level level, Char3D pattern, BlockState ritualBlock) {
+    public @Nullable BlockState matchTransformed(BlockPos origin, Level level, Char3D pattern, BlockState ritualBlock, boolean destroy) {
         List<BlockPos> inputPositions = new LinkedList<>();
         BlockState inputBlock = null;
         for (int x = 0; x < pattern.maxX(); x++) {
@@ -125,8 +125,10 @@ public class RitualStructure implements Recipe<RitualStructureContainer> {
                 }
             }
         }
-        for (BlockPos pos : inputPositions) {
-            level.destroyBlock(pos, false);
+        if (destroy) {
+            for (BlockPos pos : inputPositions) {
+                level.destroyBlock(pos, false);
+            }
         }
         return inputBlock;
     }
@@ -134,7 +136,18 @@ public class RitualStructure implements Recipe<RitualStructureContainer> {
     public Optional<BlockState> matches(BlockPos ritualPos, Level level) {
         BlockState ritualBlock = level.getBlockState(ritualPos);
         for (Char3DWithMark pattern : patterns) {
-            BlockState inputBlock = matchTransformed(pattern.getOriginInWorld(ritualPos), level, pattern, ritualBlock);
+            BlockState inputBlock = matchTransformed(pattern.getOriginInWorld(ritualPos), level, pattern, ritualBlock, false);
+            if (inputBlock != null) {
+                return Optional.of(inputBlock);
+            }
+        }
+        return Optional.empty();
+    }
+
+    public Optional<BlockState> matchesAndDestroy(BlockPos ritualPos, Level level) {
+        BlockState ritualBlock = level.getBlockState(ritualPos);
+        for (Char3DWithMark pattern : patterns) {
+            BlockState inputBlock = matchTransformed(pattern.getOriginInWorld(ritualPos), level, pattern, ritualBlock, true);
             if (inputBlock != null) {
                 return Optional.of(inputBlock);
             }
